Renamed example to sample and moved it to networks
