+++ pkg/loading/loader.go
@@ -1,8 +1,10 @@
 package loading
 
+import "iio/pkg/networks"
+
 // High-level abstraction describing a data fetcher which
 // downloads info from the specified data source and splits
 // it into training and test sets.
 type Loader interface {
-	Load() ([]*Example, []*Example, error)
+	Load() ([]*networks.Sample, []*networks.Sample, error)
 }
